+++ doc/changelog.md
@@ -1,6 +1,7 @@
 # ChangeLog
 
 * **0.11.5-M1**
+  - Add a new parameter 'docker.follow' which makes a `docker:start` blocking until the CTRL-C is pressed (#176)
   - Add a `user` parameter to the assembly configuration so that the added files are created for this user 
     (#53, #175) 
   - Fix problem when creating intermediate archive for collecting assembly files introduced with #139. The